step 1b

+++ Core/include/Acts/Propagator/detail/PointwiseMaterialInteraction.hpp
@@ -35,13 +35,13 @@ struct PointwiseMaterialInteraction {
   /// The particle direction at the interaction.
   const Vector3 dir = Vector3(0., 0., 0);
   /// The particle q/p at the interaction
-  const float qOverP = 0.0;
+  const double qOverP = 0.0;
   /// The absolute particle charge
-  const float absQ = 0.0;
+  const double absQ = 0.0;
   /// The particle momentum at the interaction
-  const float momentum = 0.0;
+  const double momentum = 0.0;
   /// The particle mass
-  const float mass = 0.0;
+  const double mass = 0.0;
   /// The particle absolute pdg
   const PdgParticle absPdg = PdgParticle::eInvalid;
   /// The covariance transport decision at the interaction
